requirements.txt
pytest==7.1.3
pyteal==0.18.1
beaker-pyteal==0.2.0
py-algorand-sdk==1.18.0